+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Use spec: https://common-changelog.org/
 
 ### Added
 
-- `zmx [hi]story` command which prints the session scrollback as plain text
+- New command `zmx [hi]story <name>` which prints the session scrollback as plain text
+- New command `zmx [r]un <name> <cmd>...` which sends a command without attaching, creating session if needed
 - Use `XDG_RUNTIME_DIR` environment variable for socket directory (takes precedence over `TMPDIR` and `/tmp`)
 
 ### Changed